chore: create refactored payment-proof-upload file structure
